Added Ability to salt and hash password; Added ability to create user; Added base database Info
