+++ pizzaMaker/Assets/Scripts/CreateUser.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 67b78c159e4f32c47a1793074abf4d60
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
