+++ pizzaMaker/Assets/Scripts/Database/DBConnection.cs
@@ -1,0 +1,51 @@
+﻿using MySql.Data.MySqlClient;
+using System;
+using UnityEngine;
+
+public class DBConnection : MonoBehaviour
+{
+    private DBConnection()
+    {
+    }
+
+    private string databaseName = string.Empty;
+    public string DatabaseName
+    {
+        get { return databaseName; }
+        set { databaseName = value; }
+    }
+
+    public string Password { get; set; }
+    private MySqlConnection connection = null;
+    public MySqlConnection Connection
+    {
+        get { return connection; }
+    }
+
+    private static DBConnection _instance = null;
+    public static DBConnection Instance()
+    {
+        if (_instance == null)
+            _instance = new DBConnection();
+        return _instance;
+    }
+
+    public bool IsConnect()
+    {
+        if (Connection == null)
+        {
+            if (String.IsNullOrEmpty(databaseName))
+                return false;
+            string connstring = string.Format("Server=13.59.229.139; database={0}; UID=googleai; password=!csc631!", databaseName);
+            connection = new MySqlConnection(connstring);
+            connection.Open();
+        }
+
+        return true;
+    }
+
+    public void Close()
+    {
+        connection.Close();
+    }
+}
