+++ pizzaMaker/Assets/Scripts/Database/DatabaseProtocol.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 698478be68c1ecf40b27f650d9e1a6db
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
