+++ pizzaMaker/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Google.Protobuf" version="3.5.1" targetFramework="net471" />
+  <package id="MySql.Data" version="8.0.15" targetFramework="net471" />
+</packages>
